feat: optmize project's folder structure
