+++ src/modules/doctor/entities/doctor.entity.ts
