+++ src/modules/doctor/repositories/DoctorRepository.ts
@@ -1,0 +1,4 @@
+import { Repository } from 'typeorm';
+import { DoctorEntity } from '../entities/doctor.entity';
+
+export class DoctorsRepository extends Repository<DoctorEntity> {}
