+++ src/modules/doctor/services/delete-doctor.service.ts
@@ -1,13 +1,13 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
-import { DoctorEntity } from './doctor.entity';
+import { DoctorEntity } from '../entities/doctor.entity';
+import { DoctorsRepository } from '../repositories/DoctorRepository';
 
 @Injectable()
 export class DeleteDoctorService {
   constructor(
     @InjectRepository(DoctorEntity)
-    private readonly doctorRepository: Repository<DoctorEntity>,
+    private readonly doctorRepository: DoctorsRepository,
   ) {}
 
   async delete(id: string): Promise<void> {
